requirements.txt
aioredis==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
certifi==2024.2.2
cffi==1.16.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==3.0.2
Django==5.0.3
django-cors-headers==4.3.1
django-environ==0.11.2
django-extensions==3.2.3
django-oauth-toolkit==2.3.0
djangorestframework==3.14.0
hiredis==2.3.2
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jwcrypto==1.5.5
loguru==0.7.2
msgpack==1.0.8
numpy==1.26.4
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pyOpenSSL==24.0.0
pytz==2024.1
requests==2.31.0
service-identity==24.1.0
setuptools==69.1.1
six==1.16.0
sqlparse==0.4.4
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
win32-setctime==1.1.0
zope.interface==6.2
